@@ -1,6 +1,6 @@
 class CategoriesController < ApplicationController
 	before_action :find_category, only: [:edit, :update, :destroy]
-	before_filter :authorize_admin
+	before_action :authorize_admin
 
 	def index
 		@category = Category.all.order('created_at ASC')
